+++ extended_mypy_django_plugin/plugin/_config.py
@@ -33,10 +33,10 @@ class Config(DjangoPluginConfig):
         "django_settings_module",
         "strict_settings",
         "scratch_path",
-        "installed_apps_script",
+        "determine_django_state_script",
     )
     scratch_path: Path
-    installed_apps_script: Path | None
+    determine_django_state_script: Path | None
 
     def parse_toml_file(self, filepath: pathlib.Path) -> None:
         toml_exit: Callable[[str], NoReturn] = partial(exit_with_error, is_toml=True)
@@ -67,11 +67,13 @@ class Config(DjangoPluginConfig):
         self.scratch_path.mkdir(parents=True, exist_ok=True)
 
         if "installed_apps_path" in config:
-            if not isinstance(config["installed_apps_script"], str):
-                toml_exit("invalid 'installed_apps_script': the setting must be a string")
-            self.installed_apps_script = filepath.parent / config["installed_apps_script"]
+            if not isinstance(config["determine_django_state_script"], str):
+                toml_exit("invalid 'determine_django_state_script': the setting must be a string")
+            self.determine_django_state_script = (
+                filepath.parent / config["determine_django_state_script"]
+            )
         else:
-            self.installed_apps_script = None
+            self.determine_django_state_script = None
 
         self.strict_settings = config.get("strict_settings", True)
         if not isinstance(self.strict_settings, bool):
@@ -100,12 +102,12 @@ class Config(DjangoPluginConfig):
         self.scratch_path = filepath.parent / parser.get(section, "scratch_path").strip("'\"")
         self.scratch_path.mkdir(parents=True, exist_ok=True)
 
-        if parser.has_option(section, "installed_apps_script"):
-            self.installed_apps_script = filepath.parent / parser.get(
-                section, "installed_apps_script"
+        if parser.has_option(section, "determine_django_state_script"):
+            self.determine_django_state_script = filepath.parent / parser.get(
+                section, "determine_django_state_script"
             ).strip("'\"")
         else:
-            self.installed_apps_script = None
+            self.determine_django_state_script = None
 
         try:
             self.strict_settings = parser.getboolean(section, "strict_settings", fallback=True)
@@ -117,8 +119,10 @@ class Config(DjangoPluginConfig):
         return {
             "django_settings_module": self.django_settings_module,
             "scratch_path": str(self.scratch_path),
-            "installed_apps_script": (
-                None if self.installed_apps_script is None else str(self.installed_apps_script)
+            "determine_django_state_script": (
+                None
+                if self.determine_django_state_script is None
+                else str(self.determine_django_state_script)
             ),
             "strict_settings": self.strict_settings,
         }