+++ docs/api/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 0.5.1 - TBD
     * Providing a return code of 2 from the installed_apps script will make dmypy not
       change version to cause a restart.
+    * Changed the ``get_installed_apps`` setting to be ``determine_django_state``
 
 .. _release-0.5.0: